\name{phase2.data}
\alias{phase2.data}
\alias{phase2.data}
\docType{package}
\title{
Prepare data for phase2 analysis
}
\description{
The package \pkg{phase2 data} consists of a collection of functions for preparing data for phase2 analysis
}
\details{
\tabular{ll}{
Package: \tab phase2.data\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2020-8-19\cr
License: GPL>=2 \cr
}
}

\author{
Chuan Hong \cr
Maintainer: Chuan Hong <chuan_hong@hms.harvard.edu>
}

\keyword{package}

